backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
google-generativeai==0.3.2
werkzeug==3.0.1
requests==2.31.0
